Refactor `chat` endpoint controller, service, and other related files.

+++ src/chat/chat.controller.ts
@@ -1,55 +1,43 @@
 import {
-  Body,
   Controller,
-  Delete,
   Get,
-  Param,
-  ParseIntPipe,
   Post,
+  Body,
+  Param,
+  Delete,
   Put,
   Query,
   UseGuards,
+  ParseIntPipe,
   UseInterceptors,
   ClassSerializerInterceptor,
+  NotFoundException,
 } from '@nestjs/common';
 import { ChatService } from './chat.service';
-import { CreateChatDTO } from './dto/create-chat.dto';
-import { UpdateChatDTO } from './dto/update-chat.dto';
+import { Chat } from './chat.entity';
+import { CreateChatDTO } from './chat-create.dto';
+import { UpdateChatDTO } from './chat-update.dto';
 import { ApiKeyEntity } from 'src/decorators/api-key.decorator';
 import { ApiKeyGuard } from 'src/guards/api-key.guard';
 import { RolesGuard } from 'src/guards/roles.guard';
 import { Roles } from 'src/decorators/roles.decorator';
-import { ApiKey, AppRoles } from 'src/api-key/api-key.entity';
-import { ChatResponseDTO } from './dto/chat-response.dto';
+import { ChatResponseDTO } from './chat-response.dto';
 import {
   ApiTags,
   ApiOperation,
-  ApiQuery,
-  ApiResponse,
-  ApiBody,
   ApiParam,
+  ApiBody,
   ApiSecurity,
+  ApiQuery,
 } from '@nestjs/swagger';
-import { ErrorResponseDTO } from 'src/dto/error-response.dto';
 import { ApiOkResponseWithWrapper } from 'src/decorators/api-ok-response-wrapper.decorator';
+import { FindChatsQueryDTO } from './chat-find-query.dto';
+import { FindChatsResponseDTO } from './chat-find-response.dto';
+import { ApiKey, AppRoles } from 'src/api-key/api-key.entity';
+import { CommonApiResponses } from 'src/decorators/common-api-responses.decorator';
 
 @ApiTags('Chats')
-@ApiResponse({
-  status: 400,
-  description: 'Bad Request',
-  type: ErrorResponseDTO,
-})
-@ApiResponse({
-  status: 401,
-  description: 'Unauthorized',
-  type: ErrorResponseDTO,
-})
-@ApiResponse({ status: 404, description: 'Not Found', type: ErrorResponseDTO })
-@ApiResponse({
-  status: 500,
-  description: 'Internal Server Error',
-  type: ErrorResponseDTO,
-})
+@CommonApiResponses()
 @ApiSecurity('apiKey')
 @Controller('chats')
 @UseGuards(ApiKeyGuard, RolesGuard)
@@ -61,32 +49,34 @@ export class ChatController {
   @Roles(AppRoles.ADMIN)
   @ApiOperation({ summary: 'Retrieve a list of chats' })
   @ApiQuery({
-    name: 'search',
-    required: false,
-    description: 'Search filter for chats',
-  })
-  @ApiQuery({
-    name: 'limit',
-    required: false,
-    description: 'Limit the number of results',
-  })
-  @ApiQuery({
-    name: 'offset',
-    required: false,
-    description: 'Offset for pagination',
+    type: FindChatsQueryDTO,
   })
   @ApiOkResponseWithWrapper({
-    description: 'List of chats',
+    description: 'List of chats along with pagination details and filters',
     status: 200,
-    type: ChatResponseDTO,
-    isArray: true,
+    type: FindChatsResponseDTO,
   })
   async findAll(
-    @Query('search') search?: string,
-    @Query('limit', new ParseIntPipe({ optional: true })) limit?: number,
-    @Query('offset', new ParseIntPipe({ optional: true })) offset?: number,
-  ): Promise<ChatResponseDTO[]> {
-    return await this.chatService.findAll(search, limit, offset);
+    @Query() query: FindChatsQueryDTO,
+  ): Promise<FindChatsResponseDTO> {
+    const { limit, offset, search, apiKeyId, assistantName } = query;
+
+    const chats: Chat[] = await this.chatService.findAll(
+      limit,
+      offset,
+      search,
+      apiKeyId,
+      assistantName,
+    );
+
+    return {
+      limit,
+      offset,
+      search,
+      apiKeyId,
+      assistantName,
+      data: chats,
+    };
   }
 
   @Get(':chatId')
@@ -99,9 +89,13 @@ export class ChatController {
     type: ChatResponseDTO,
   })
   async findOne(
-    @Param('chatId', new ParseIntPipe({ optional: true })) chatId: number,
+    @Param('chatId', ParseIntPipe) chatId: number,
   ): Promise<ChatResponseDTO> {
-    return await this.chatService.findOne(chatId);
+    const chat = await this.chatService.findOne(chatId);
+    if (!chat) {
+      throw new NotFoundException(`Chat with ID ${chatId} not found`);
+    }
+    return chat;
   }
 
   @Post()
@@ -117,7 +111,7 @@ export class ChatController {
     @Body() createChatDto: CreateChatDTO,
     @ApiKeyEntity() apiKey: ApiKey,
   ): Promise<ChatResponseDTO> {
-    return await this.chatService.create(apiKey.id, createChatDto);
+    return this.chatService.create(apiKey.id, createChatDto);
   }
 
   @Put(':chatId')
@@ -131,10 +125,14 @@ export class ChatController {
     type: ChatResponseDTO,
   })
   async update(
-    @Param('chatId', new ParseIntPipe()) chatId: number,
+    @Param('chatId', ParseIntPipe) chatId: number,
     @Body() updateChatDto: UpdateChatDTO,
   ): Promise<ChatResponseDTO> {
-    return await this.chatService.update(chatId, updateChatDto);
+    const updatedChat = await this.chatService.update(chatId, updateChatDto);
+    if (!updatedChat) {
+      throw new NotFoundException(`Chat with ID ${chatId} not found`);
+    }
+    return updatedChat;
   }
 
   @Delete(':chatId')
@@ -146,8 +144,17 @@ export class ChatController {
     status: 200,
   })
   async delete(
-    @Param('chatId', new ParseIntPipe()) chatId: number,
-  ): Promise<void> {
+    @Param('chatId', ParseIntPipe) chatId: number,
+  ): Promise<{ statusCode: number; message: string }> {
+    const chat = await this.chatService.findOne(chatId);
+    if (!chat) {
+      throw new NotFoundException(`Chat with ID ${chatId} not found`);
+    }
     await this.chatService.delete(chatId);
+
+    return {
+      statusCode: 200,
+      message: 'Chat deleted successfully',
+    };
   }
 }
